+++ modules-wip/raster-layers/src/shadermodules/spectral-indices/evi.ts
@@ -11,7 +11,7 @@ import {ShaderModule} from '../types';
 // green: Landsat 8 band 4
 // blue: Landsat 8 band 2
 //
-// EVI = 2.5 * ((Band 5 – Band 4) / (Band 5 + 6 * Band 4 – 7.5 * Band 2 + 1))
+// EVI = 2.5 * ((Band 5 - Band 4) / (Band 5 + 6 * Band 4 - 7.5 * Band 2 + 1))
 // https://www.usgs.gov/land-resources/nli/landsat/landsat-enhanced-vegetation-index
 const fs = /* glsl */ `\
 float enhanced_vegetation_index_calc(vec4 image) {