+++ modules-wip/raster-layers/src/shadermodules/spectral-indices/savi.ts
@@ -6,7 +6,7 @@
 import {ShaderModule} from '../types';
 
 // Calculate soil-adjusted vegetation index
-// SAVI = ((Band 5 – Band 4) / (Band 5 + Band 4 + 0.5)) * (1.5).
+// SAVI = ((Band 5 - Band 4) / (Band 5 + Band 4 + 0.5)) * (1.5).
 // https://www.usgs.gov/land-resources/nli/landsat/landsat-soil-adjusted-vegetation-index
 const fs = /* glsl */ `\
 float soil_adjusted_vegetation_index_calc(vec4 image) {