+++ modules/graph-layers/src/layouts/gpu-force/gpu-force-layout.ts
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
-import {GraphLayout, GraphLayoutOptions} from '../../core/graph-layout';
+import {GraphLayout, GraphLayoutProps} from '../../core/graph-layout';
 
-export type GPUForceLayoutOptions = GraphLayoutOptions & {
+export type GPUForceLayoutOptions = GraphLayoutProps & {
   alpha?: number;
   resumeAlpha?: number;
   nBodyStrength?: number;
@@ -17,7 +17,7 @@ export type GPUForceLayoutOptions = GraphLayoutOptions & {
  * @todo this layout should be updated with the organizational and logic improvements made in d3-force
  */
 export class GPUForceLayout extends GraphLayout<GPUForceLayoutOptions> {
-  static defaultOptions: Required<GPUForceLayoutOptions> = {
+  static defaultProps: Required<GPUForceLayoutOptions> = {
     alpha: 0.3,
     resumeAlpha: 0.1,
     nBodyStrength: -900,
@@ -35,15 +35,15 @@ export class GPUForceLayout extends GraphLayout<GPUForceLayoutOptions> {
   private _callbacks: any;
 
   constructor(options: GPUForceLayoutOptions = {}) {
-    const _options = {
-      ...GPUForceLayout.defaultOptions,
+    const props = {
+      ...GPUForceLayout.defaultProps,
       ...options
     };
 
-    super(_options);
+    super(props);
 
     this._name = 'GPU';
-    this._options = _options;
+    this.props = props;
     // store graph and prepare internal data
     this._d3Graph = {nodes: [], edges: []};
     this._nodeMap = {};
@@ -109,7 +109,7 @@ export class GPUForceLayout extends GraphLayout<GPUForceLayoutOptions> {
 
     this._worker = new Worker(new URL('./worker.js', import.meta.url).href);
     const {alpha, nBodyStrength, nBodyDistanceMin, nBodyDistanceMax, getCollisionRadius} =
-      this._options;
+      this.props;
     this._worker.postMessage({
       nodes: this._d3Graph.nodes,
       edges: this._d3Graph.edges,